% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PneumatronExperiments.R
\name{add_experiments_from_df}
\alias{add_experiments_from_df}
\title{Add Experiments from Data Frame}
\usage{
add_experiments_from_df(experiments_table, new_experiments_df)
}
\arguments{
\item{experiments_table}{Existing table of experiments.}

\item{new_experiments_df}{Data frame with new experiments.}
}
\value{
The updated table of experiments.
}
\description{
Add experiments from a data frame to the existing table.
}
