% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PneumatronExperiments.R
\name{add_new_experiments}
\alias{add_new_experiments}
\title{Add New Experiments}
\usage{
add_new_experiments(
  experiments_table,
  pneumatron_file,
  water_potential_file,
  pneumatron_id,
  begin,
  end,
  min_datetime = NA,
  max_datetime = NA,
  min_gas_discharge = NA,
  max_gas_discharge = NA,
  max_wp = NA,
  obs = NA
)
}
\arguments{
\item{experiments_table}{Existing table of experiments.}

\item{pneumatron_file}{Path to the Pneumatron file.}

\item{water_potential_file}{Path to the water potential file.}

\item{pneumatron_id}{Device ID.}

\item{begin}{Start datetime of the experiment.}

\item{end}{End datetime of the experiment.}

\item{min_datetime}{(Optional) Minimum datetime.}

\item{max_datetime}{(Optional) Maximum datetime.}

\item{min_gas_discharge}{(Optional) Minimum gas discharge.}

\item{max_gas_discharge}{(Optional) Maximum gas discharge.}

\item{max_wp}{(Optional) Minimum water potential.}

\item{obs}{(Optional) Additional experiment observation.}
}
\value{
The updated table of experiments.
}
\description{
Add new experiments to the existing table.
}
