% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PneumatronExperiments.R
\name{as_pneumatron_experiment}
\alias{as_pneumatron_experiment}
\title{Convert Data Frame to PneumatronExperiments}
\usage{
as_pneumatron_experiment(df)
}
\arguments{
\item{df}{Data frame to be converted.}
}
\value{
An object of class PneumatronExperiments.
}
\description{
User-facing function to convert a data frame into PneumatronExperiments.
}
