% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateGasDischarged.R
\name{calculateGasDischarged}
\alias{calculateGasDischarged}
\alias{calculateGasDischarged,PneumatronDatabase-method}
\title{Calculate gas discharge from Pneumatron output}
\usage{
calculateGasDischarged(
  object,
  pressure_interval = 15,
  reservoir = 2.6,
  p_atm = 101.3,
  fixed_p_atm = TRUE,
  temp = 25,
  fixed_temp = TRUE,
  R = 8.3144621
)

\S4method{calculateGasDischarged}{PneumatronDatabase}(
  object,
  pressure_interval = 15,
  reservoir = 2.6,
  p_atm = 101.3,
  fixed_p_atm = TRUE,
  temp = 25,
  fixed_temp = TRUE,
  R = 8.3144621
)
}
\arguments{
\item{object}{A PneumatronDatabase object.}

\item{pressure_interval}{Interval to calculate pressure difference in seconds. Defauts 15s.}

\item{reservoir}{Tubing volume of the reservoir in mL.}

\item{p_atm}{Fixed atmospheric pressure in kPa, used if fixed_p_atm is TRUE.}

\item{fixed_p_atm}{Logical, if TRUE uses a fixed atmospheric pressure for calculations, otherwise uses pressure from data.}

\item{temp}{Fixed temperature in Celsius, used if fixed_temp is TRUE.}

\item{fixed_temp}{Logical, if TRUE uses a fixed temperature for calculations, otherwise uses temperature from data. Defauts is TRUE.}

\item{R}{Ideal gas law constant.}
}
\value{
A GasDischarged object with calculated gas discharge values.
}
\description{
This method calculates the gas discharge from the Pneumatron database.
}
\examples{
\dontrun{
database <- open_pneumatron_database("path/to/your/data.csv")
gas_discharged <- calculateGasDischarged(database)
}
}
