% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PneumatronExperiments.R
\name{check_experiment_table_classes}
\alias{check_experiment_table_classes}
\title{Check Column Classes}
\usage{
check_experiment_table_classes(df)
}
\arguments{
\item{df}{Data frame whose columns are to be checked.}
}
\value{
TRUE if all columns match the expected classes, otherwise stops with an error.
}
\description{
Check if the columns in the data frame have the expected classes.
}
