% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PneumatronExperiments.R
\name{convert_to_pneumatron_experiment}
\alias{convert_to_pneumatron_experiment}
\title{Convert Data Frame to PneumatronExperiments}
\usage{
convert_to_pneumatron_experiment(df, start_id)
}
\arguments{
\item{df}{Data frame to be converted.}

\item{start_id}{Initial ID for the experiments.}
}
\value{
An object of class PneumatronExperiments.
}
\description{
Internal function to convert a data frame to PneumatronExperiments, adding ID if necessary.
}
