% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_pneumatron_data.R
\name{detect_data_format}
\alias{detect_data_format}
\title{Detect the Format of Pneumatron Data File}
\usage{
detect_data_format(file_path)
}
\arguments{
\item{file_path}{The path to the CSV file whose format needs to be detected.}
}
\value{
A character string indicating the detected data format ('V2', 'V2-update', 'V3', 'V3-old', 'V4').
If the format cannot be detected or is unsupported, the function will stop with an error message.
}
\description{
Automatically detects the format of a Pneumatron data file by analyzing the header of the file.
The function reads only the first row to determine the number of columns and checks for specific headers
to infer the format version of the file.
}
\examples{
format <- detect_data_format("path/to/your/datafile.csv")
}
