% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PneumatronExperiments.R
\name{extract_experiment}
\alias{extract_experiment}
\title{Extract a Single Experiment}
\usage{
extract_experiment(experiments_table, experiment_id)
}
\arguments{
\item{experiments_table}{Existing table of experiments.}

\item{experiment_id}{ID of the experiment to be extracted.}
}
\value{
The single experiments.
}
\description{
Extract a single experiment from experiments table.
}
