% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PneumatronData.R
\name{get_data}
\alias{get_data}
\alias{get_data,PneumatronData-method}
\title{Get data from PneumatronData object}
\usage{
get_data(x)

\S4method{get_data}{PneumatronData}(x)
}
\arguments{
\item{x}{A PneumatronData object.}
}
\value{
A data.table containing the data stored in the PneumatronData object.
}
\description{
This method retrieves the data stored in a PneumatronData object.
}
\examples{
\dontrun{
pneumatron_data <- PneumatronData(data = df)
data <- get_data(pneumatron_data)
}
}
