% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_experiments_table.R
\name{is_experiments_table}
\alias{is_experiments_table}
\title{Check if a Data Frame is in Pneumatron Experiments Format}
\usage{
is_experiments_table(x)
}
\arguments{
\item{x}{A data frame to be checked.}
}
\value{
A logical value indicating whether the data frame is in the Pneumatron experiments format.
}
\description{
This function checks if a given data frame conforms to the expected format for Pneumatron experiments.
}
\details{
The function verifies that the input data frame has the required columns:
"id", "pneumatron_id", "start_datetime", "end_datetime", and "pneumatron_file".
Additionally, it checks that:
\itemize{
\item "id" is numeric
\item "pneumatron_id" is numeric
\item "start_datetime" is of class POSIXct
\item "end_datetime" is of class POSIXct
\item "pneumatron_file" is a character vector
}
}
\examples{
\dontrun{
df <- data.frame(
  id = 1:5,
  pneumatron_id = 1:5,
  start_datetime = as.POSIXct(Sys.time() + 1:5),
  end_datetime = as.POSIXct(Sys.time() + 6:10),
  pneumatron_file = c("file1.csv", "file2.csv", "file3.csv", "file4.csv", "file5.csv")
)
is_experiments_table(df) # returns TRUE
}
}
