% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_pneumatron_data.R
\name{open_pneumatron_data}
\alias{open_pneumatron_data}
\title{Open Pneumatron Data from File}
\usage{
open_pneumatron_data(file_path, data_format = "V4")
}
\arguments{
\item{file_path}{The path to the CSV file.}

\item{data_format}{The format of the data file ('V2', 'V2-update', 'V3', 'V3-old', 'V4').
Defaults to 'V4'.}
}
\value{
An object of class PneumatronData.
}
\description{
Reads Pneumatron measurement data from a CSV file using the specified data format.
Defaults to format V4 if no format is specified.
}
\examples{
\dontrun{
pneumatron_data <- open_pneumatron_data("path/to/your/datafile.csv", data_format = "V4")
}
}
