% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_pneumatron_data.R
\name{open_pneumatron_mult_database}
\alias{open_pneumatron_mult_database}
\title{Open Mult-Pneumatron Data from File}
\usage{
open_pneumatron_mult_database(
  files_path,
  ids,
  vld = FALSE,
  datetime_format = "\%y/\%m/\%d \%H:\%M"
)
}
\arguments{
\item{files_path}{A character vector containing the paths to the CSV files.}

\item{ids}{A numeric vector of 'pneumatron id' corresponding to each file. Must be the same length as files_path.}

\item{vld}{Logical indicating if you are working with VLD. Default is FALSE.}

\item{datetime_format}{A character string specifying the date-time format to be passed to as.POSIXct. Default is "\%y/\%m/\%d \%H:\%M".}
}
\value{
An object of class PneumatronDatabase.
}
\description{
Reads Pneumatron measurement data from Mult-Pneumatron file format and
converts it to the standard format of this package.
}
\details{
The 'pneumatron id' specified in ids will be used to merge water potential data.
}
\examples{
\dontrun{
database <- open_pneumatron_mult_database(c("sample_1.csv", "sample_2.csv"), ids = c(1,2))
}
}
