% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PneumatronData.R
\docType{class}
\name{PneumatronData-class}
\alias{PneumatronData-class}
\alias{PneumatronData}
\title{Class PneumatronData}
\description{
This class is designed for managing large datasets from Pneumatron devices efficiently.
It utilizes \code{data.table} for data manipulation, ensuring fast processing and easy handling of data.
}
\section{Slots}{

\describe{
\item{\code{file_path}}{Character vector, represents the file path for the data file.}

\item{\code{data_format}}{Character vector, specifies the Pneumatron data format of the data (e.g., V2, V3, V4).}
}}

\examples{
# To convert a data.frame to PneumatronData, first ensure your data.frame includes
# the required columns: id, measure, log_line, and pressure. Here's an example:
df <- data.frame(
  id = c(1, 1, 1, 1, 1),
  measure = c(1, 2, 3, 4, 5),
  log_line = c(1, 2, 3, 4, 5),
  pressure = c(100, 40, 40.7, 41.5, 41.9),
  datetime = as.POSIXct(c("2024-05-13 13:54:26", "2024-05-13 13:54:26", "2024-05-13 13:54:27", "2024-05-13 13:54:27", "2024-05-13 13:54:28"))
)
# Now, convert the data.frame to a PneumatronData object.
pneumatron_data <- PneumatronData(df)
print(pneumatron_data)

}
