% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PneumatronData.R
\docType{class}
\name{PneumatronData-class}
\alias{PneumatronData-class}
\alias{PneumatronData}
\title{Class PneumatronData}
\usage{
PneumatronData(data, file_path = NA_character_, data_format = NA_character_)
}
\arguments{
\item{data}{data.table, represents the data stored in a data.table.}

\item{file_path}{Character vector, represents the file path for the data file.}

\item{data_format}{Character vector, specifies the Pneumatron data format of the data (e.g., V2, V3, V4).}
}
\description{
This class is designed for managing large datasets from Pneumatron devices efficiently.
It utilizes \code{data.table} for data manipulation, ensuring fast processing and easy handling of data.
}
\section{Slots}{

\describe{
\item{\code{data}}{data.table, represents the data stored in a data.table.}

\item{\code{file_path}}{Character vector, represents the file path for the data file.}

\item{\code{data_format}}{Character vector, specifies the Pneumatron data format of the data (e.g., V2, V3, V4).}
}}

