% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateGasDischarged.R
\name{prepare_data_for_gas_discharged}
\alias{prepare_data_for_gas_discharged}
\title{Prepare data for gas discharged calculations}
\usage{
prepare_data_for_gas_discharged(data, pressure_interval)
}
\arguments{
\item{data}{Pneumatron output object, or a data frame with properly named columns.}

\item{pressure_interval}{Final pressure time in seconds.}
}
\value{
A data table prepared for gas discharge calculations.
}
\description{
This function is used internally for preparing data for gas discharge calculations. It is not exported for end user.
It prepares and formats the data by filtering and calculating necessary parameters for the discharge calculation.
}
