% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PneumatronData_summary.R
\name{summary,PneumatronData-method}
\alias{summary,PneumatronData-method}
\title{Summary method for PneumatronData}
\usage{
\S4method{summary}{PneumatronData}(object)
}
\arguments{
\item{object}{A PneumatronData object.}
}
\description{
Provides a summary of the PneumatronData object, including the number of measurements,
start and end times for each Pneumatron device, and identifies measurements that need
to be checked or removed.
}
\examples{
\dontrun{
pneumatron_data <- PneumatronData(data = df)
summary(pneumatron_data)
}
}
