% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PneumatronExperiments.R
\name{update_experiment}
\alias{update_experiment}
\title{Update an Experiment}
\usage{
update_experiment(
  experiments_table,
  experiment_id,
  pneumatron_file = NA,
  water_potential_file = NA,
  pneumatron_id = NA,
  begin = NA,
  end = NA,
  min_datetime = NA,
  max_datetime = NA,
  min_gas_discharge = NA,
  max_gas_discharge = NA,
  max_wp = NA,
  obs = NA
)
}
\arguments{
\item{experiments_table}{Existing table of experiments.}

\item{experiment_id}{ID of the experiment to be updated.}

\item{pneumatron_file}{(Optional) Path to the Pneumatron file.}

\item{water_potential_file}{(Optional) Path to the water potential file.}

\item{pneumatron_id}{(Optional) Device ID.}

\item{begin}{(Optional) Start datetime of the experiment.}

\item{end}{(Optional) End datetime of the experiment.}

\item{min_datetime}{(Optional) Minimum datetime.}

\item{max_datetime}{(Optional) Maximum datetime.}

\item{min_gas_discharge}{(Optional) Minimum gas discharge.}

\item{max_gas_discharge}{(Optional) Maximum gas discharge.}

\item{max_wp}{(Optional) Minimum water potential.}

\item{obs}{(Optional) Additional experiment observation.}
}
\value{
The updated table of experiments.
}
\description{
Update an existing experiment in the table.
}
