% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_PneumatronData.R
\name{get_data}
\alias{get_data}
\alias{get_data,PneumatronDatabase-method}
\title{Get data from PneumatronDatabase object}
\usage{
get_data(x)

\S4method{get_data}{PneumatronDatabase}(x)
}
\arguments{
\item{x}{A PneumatronDatabase object.}
}
\value{
A data.table containing the data stored in the PneumatronDatabase object.
}
\description{
This method retrieves the data stored in a PneumatronDatabase object.
}
\examples{
\dontrun{
database <- open_pneumatron_database("path/to/your/datafile.csv")
data <- get_data(database)
}
}
