% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_pneumatron_data.R
\name{open_pneumatron_mult_database}
\alias{open_pneumatron_mult_database}
\title{Open Mult-Pneumatron Data from File}
\usage{
open_pneumatron_mult_database(
  files_path,
  ids,
  vld = FALSE,
  datetime_format = "\%y/\%m/\%d \%H:\%M"
)
}
\arguments{
\item{files_path}{The path to the CSV files.}

\item{ids}{The 'pneumatron id' of each file. Must be the same size as file_path.}

\item{vld}{if you are working with VLD mark as TRUE. Defaut is FALSE}

\item{datetime_format}{It will be passed for \code{as.POSIXct} format. Default = "\%y/\%m/\%d \%H:\%M"}
}
\value{
An object of class PneumatronDatabase.
}
\description{
Reads Pneumatron measurement data from Mult-Pneumatron file format and
convert to the standard format of this package.
}
\details{
The 'pneumatron id' specified in 'ids' will be used to merge water potential data.
}
\examples{
\dontrun{
database <- open_pneumatron_mult_database(c("sample_1.csv", "sample_2.csv"), ids = c(1,2))
}
}
