% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PneumatronData_reportUptime.R
\name{reportUptime}
\alias{reportUptime}
\title{Report Uptime for Pneumatron Devices}
\usage{
reportUptime(object, lower_interval = 750, upper_interval = 1050)
}
\arguments{
\item{object}{A PneumatronData object.}

\item{lower_interval}{The expected interval between measurements in seconds.
Defaults to 780 seconds (13 min).}

\item{upper_interval}{The expected interval between measurements in seconds.
Defaults to 1020 seconds (17 min).}
}
\value{
A data.table with the uptime information for each device.
}
\description{
A generic function to reports the continuous uptime for each Pneumatron
device, considering measurements made at a specified interval (e.g. 15min).
Any gap exceeding the interval is considered as downtime.
}
\examples{
data <- open_pneumatron_data("path/to/your/datafile.csv")
uptime_summary <- reportUptime(pneumatron_data)
print(uptime_summary)
}
